Bugfix: Fix really annoying timing bug
This mainly affected HNW, where a session could be launched and then shut down before it was used.  The consequence was that, once you loaded a full session, you would actually have two event loops running with `rAF`, each one pulling events out of `Updater` and preventing the other from receiving and rendering them.

This, of course, would not happen deterministically, so it was tons of fun to figure out what was going on.